# proxy/docs/swagger.yaml
definitions:
  main.ErrorResponse:
    description: Ошибка, возникающая при обработке запроса
    properties:
      "200":
        type: string
      "400":
        type: string
      "500":
        type: string
    type: object
  main.RequestAddressSearch:
    properties:
      query:
        type: string
    type: object
  main.ResponseAddress:
    properties:
      suggestions:
        items:
          properties:
            data:
              $ref: '#/definitions/main.ResponseAddresses'
          type: object
        type: array
    type: object
  main.ResponseAddresses:
    properties:
      geo_lat:
        type: string
      geo_lon:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server Petstore server.
  title: Swagger Example API
  version: "1.0"
paths:
  /api/address/geocode:
    post:
      description: This endpoint allows you to get geo coordinates by address
      parameters:
      - description: Address search query
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/main.RequestAddressSearch'
      responses:
        "200":
          description: Успешное выполнение
          schema:
            $ref: '#/definitions/main.ResponseAddress'
        "400":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Ошибка подключения к серверу
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get Geo Coordinates
  /api/address/search:
    post:
      description: This endpoint allows you to get geo coordinates by address
      parameters:
      - description: Address search query
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/main.RequestAddressSearch'
      responses:
        "200":
          description: Успешное выполнение
          schema:
            $ref: '#/definitions/main.ResponseAddress'
        "400":
          description: Ошибка запроса
          schema:
            $ref: '#/definitions/main.ErrorResponse'
        "500":
          description: Ошибка подключения к серверу
          schema:
            $ref: '#/definitions/main.ErrorResponse'
      summary: Get Geo Coordinates
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"